Don't mush words together in TTS
This issue only appears to happen when a page ends far from a period and
doesn't end in a hyphen, but when it happens it causes the last word of
one page to be prepended to the first word of the next page and then the
TTS engine tries to read it as one word and it's terrible.

Fixed by "either drop a - at the end or add a space at the end, then
append the next chunk".